requirements.txt
aiofiles
aiohttp
apscheduler
asyncio
bing_image_downloader
bs4
deepai
edge-tts
ffmpeg-python
future
gitpython
google-api-python-client
gpytranslate
gtts
heroku3
httpx[http2]==0.28.1
kurigram==2.2.14
lexica-api
motor
nekosbest
numpy
opencv-python
pillow==12.0.0
psutil
ntgcalls==2.0.6
py-tgcalls==2.2.8
pycryptodome
pydub
pyfiglet
pyshorteners
python-dotenv
python-whois
pyyaml
qrcode
requests
speedtest-cli
SpeechRecognition
spotipy
telegraph
tgcrypto
unidecode
wget
yt-dlp[default]==2025.11.12
git+https://github.com/CertifiedCoders/youtube-search-python
